Fix compiler warning on Windows
git-svn-id: svn+ssh://svn.code.sf.net/p/virtualgl/code/trunk@3180 799e4f7b-5fd2-41f6-823c-2ecc41bc7f0b

+++ util/rrsocket.cpp
@@ -231,7 +231,7 @@ void rrsocket::connect(char *servername, unsigned short port)
 	{
 		if((_sslctx=SSL_CTX_new(SSLv23_client_method()))==NULL) _throwssl();
 		if((_ssl=SSL_new(_sslctx))==NULL) _throwssl();
-		if(!SSL_set_fd(_ssl, _sd)) _throwssl();
+		if(!SSL_set_fd(_ssl, (int)_sd)) _throwssl();
 		int ret=SSL_connect(_ssl);
 		if(ret!=1) throw(sslerror("rrsocket::connect", _ssl, ret));
 		SSL_set_connect_state(_ssl);
